requirements.txt
chromadb
requests
langchain
accelerate
langchain_community
streamlit
together
python-dotenv
pypdf
sentence-transformers